style: 에그데이 프로그래스바 넣어서 단계별로 고침 #165
