+++ src/app/(pages)/(club)/club/_components/ProgressBar.tsx
@@ -11,7 +11,7 @@ const ProgressBar = ({ currentStep, totalSteps }: ProgressBarProps) => {
       <div className="w-full h-1 bg-gray-200 rounded-full">
         <div
           className={`h-full rounded-full transition-all duration-300 ease-in-out ${
-            totalSteps === 7 ? "bg-primary-500" : "bg-primary-900"
+            totalSteps === 7 ? "bg-primary-500" : totalSteps === 4 ? "bg-primary-300" : "bg-primary-900"
           }`}
           style={{ width: `${progress}%` }}
         />
