+++ src/app/(pages)/(club)/club/_components/regularClub/Category.tsx
@@ -58,7 +58,7 @@ const Category = ({ formData, setFormData }: EggClubProps) => {
       <Text variant="header-18" className="flex items-center mb-6 h-11">
         어떤 주제로 시작해볼까요?
       </Text>
-      <div className="flex flex-col gap-2 h-[400px] overflow-y-auto scrollbar-hide">
+      <div className="flex flex-col gap-2 h-[562px] overflow-y-auto scrollbar-hide">
         {mainCategories?.map((main) => (
           <div
             key={main.main_category_id}
