+++ src/app/(pages)/(club)/club/_utils/localStorage.ts
@@ -1,3 +1,4 @@
 // 로컬스토리지 키
 export const ONETIME_CLUB_CREATE = "ONETIME_CLUB_CREATE";
 export const REGULAR_CLUB_CREATE = "REGULAR_CLUB_CREATE";
+export const NOTIFICATION_CREATE = "NOTIFICATION_CREATE";
