+++ src/types/features/club/eggday.types.ts
@@ -1,3 +1,4 @@
+import { SetStateAction } from "react";
 import { Tables } from "../../core/database.types";
 
 // 정기적모임 안의 공지 인풋 타입
@@ -20,4 +21,14 @@ export type EggDayMemberRequired = Pick<Tables<"egg_day_member">, "egg_day_id" |
 
 export type EggDayWithEggDayMember = EggDay & {
   egg_day_member: { count: number }[];
+};
+
+export interface EggDayFormWithImageFile extends Omit<EggDayRequired, "egg_day_image"> {
+  egg_day_image: File;
+}
+
+// 에그팝 폼 props
+export type EggDayProps = {
+  formData: EggDayFormWithImageFile;
+  setFormData: React.Dispatch<SetStateAction<EggDayFormWithImageFile>>;
 };
